third_party/WebKit/LayoutTests/http/tests/devtools/console/console-viewport-indices-expected.txt

Verifies viewport's visible and active message ranges.


Running: testEmptyViewport
Calculated visible range: -1 to -1, Total: 0
Actual visible range: -1 to -1, Total: 0
Active range: -1 to -1, Total: 0


Running: testFirstLastVisibleIndices
Logging 100 messages
Force item to be first: 0
Calculated visible range: 0 to 10, Total: 11
Actual visible range: 0 to 10, Total: 11
Active range: 0 to 24, Total: 25

Force item to be first: 1
Calculated visible range: 1 to 11, Total: 11
Actual visible range: 1 to 11, Total: 11
Active range: 0 to 24, Total: 25

Scroll a bit down: 15px
Calculated visible range: 1 to 12, Total: 12
Actual visible range: 1 to 12, Total: 12
Active range: 0 to 24, Total: 25

Force item to be last: 50
Calculated visible range: 40 to 50, Total: 11
Actual visible range: 40 to 50, Total: 11
Active range: 35 to 59, Total: 25

Force item to be last: 99
Calculated visible range: 89 to 99, Total: 11
Actual visible range: 89 to 99, Total: 11
Active range: 84 to 99, Total: 16
